requirements-github.txt
tensorflow>=2.15.0
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.23.5,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0
pandas>=2.0.0
scikit-learn>=1.3.0